% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teller_2016_make_data.R
\name{aggLags}
\alias{aggLags}
\title{Function from Teller 2016 (doi 10.1111/2041-210X.12486) to create distributed lags of monthly data}
\usage{
aggLags(datC, fun, meas, seg)
}
\description{
Function from Teller 2016 (doi 10.1111/2041-210X.12486) to create distributed lags of monthly data
}
