% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleData_lags_Feb18.r
\name{assemble.data.lags}
\alias{assemble.data.lags}
\title{Assemble human case data and environmental lags}
\usage{
assemble.data.lags(pop, cases, NEdat, spi, spei, target.date, start.year,
  in.seed)
}
\arguments{
\item{pop}{#**# Add documentation}

\item{cases}{#**# Add documentation}

\item{NEdat}{#**# Add documentation}

\item{spi}{#**# Add documentation}

\item{spei}{#**# Add documentation}

\item{target.date}{The last date to include for calculation of lags}

\item{start.year}{The first year to include in the training data}

\item{in.seed}{The starting number for the random number generator. This makes the results repeatable.}
}
\description{
Assemble human case data and environmental lags
}
