% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.NE.WNV.R
\name{call.flm}
\alias{call.flm}
\title{Main function}
\usage{
call.flm(
  pop,
  cases,
  weather,
  spi,
  spei,
  target.date = "2018-02-01",
  start.year = 2002,
  in.seed = NULL,
  lag.lengths = c(12, 18, 24, 30, 36),
  fillzeros = FALSE,
  nsims = 0
)
}
\arguments{
\item{pop}{County populations, a data frame with 5 variables \preformatted{County}, \preformatted{fips} (5 \strong{characters}), \preformatted{year}, \preformatted{pop100K}, \preformatted{density}.}

\item{cases}{data on annual numbers of human cases in each county. A data.frame with 3 variables, \preformatted{County}, \preformatted{year}, and \preformatted{cases}.}

\item{weather}{monthly temperature and precipitation data for for each county. A data.frame with  \preformatted{County}, \preformatted{fips} (5 \strong{characters}), \preformatted{year}, \preformatted{month} (integer), \preformatted{tmean}, and \preformatted{ppt}.}

\item{spi}{monthly values of the Standardized Precipitation Index for each county. \preformatted{County}, \preformatted{fips} (5 \strong{characters}), \preformatted{year}, \preformatted{month} (integer), \preformatted{spi}.}

\item{spei}{monthly values of the Standardized Precipitation and Evapotranspiration Index for each county. \preformatted{County}, \preformatted{fips} (5 \strong{characters}), \preformatted{year}, \preformatted{month} (integer), \preformatted{spei}.}

\item{target.date}{The last date to include for calculation of lags, a character string with ISO XXX format (yyyy-mm-dd).}

\item{start.year}{The first year to include in the training data. Should be coercible to integer.}

\item{in.seed}{If not NULL, the starting number for the random number generator. This makes the results repeatable. If NULL, treats cases as actual data}

\item{lag.lengths}{the number of months to go backwards when creating lag matrices. Numeric vector.}

\item{fillzeros}{Logical. If true, add zero predictions for counties that never had any cases.}

\item{nsims}{Integer. Number of samples to draw from posterior distribution. Defaults to zero, which has the expected value of cases in predcases.}
}
\description{
Main function
}
