% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functional.r
\name{extract_functional}
\alias{extract_functional}
\title{Extract coefficients for functional lags}
\usage{
extract_functional(x, se = TRUE, select = NULL, scale = -1)
}
\arguments{
\item{x}{is a fitted gam() with m >= 1 linear functional terms}

\item{se}{logical TRUE says to return the SE estimate as well}

\item{select}{allows restricting which functional term is returned; default is all}

\item{scale}{may be vestigial}
}
\value{
returns a data.frame with columns term, lag, coef, se
}
\description{
Extract coefficients for functional lags
}
