% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\name{flm.convert.weather.data}
\alias{flm.convert.weather.data}
\title{Convert weather.data to NEdat object}
\usage{
flm.convert.weather.data(weather.data, cofips.lookup)
}
\arguments{
\item{weather.data}{Input weather data in format district, doy, tmeanc, pr, date, plus other unused columns.}

\item{cofips.lookup}{A link between cofips and district (using the pop object, to prevent needing another input object)}
}
\value{
NEdat Output weather data in format: cofips, year, month, temp, ppt
}
\description{
Convert weather.data to NEdat object
}
