% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\docType{package}
\name{flm}
\alias{flm}
\title{Functional Linear Modeling for Vector-borne Disease}
\description{
Use functional linear modeling statistical approach to predict mosquito-borne disease.
Developed in the context of West Nile virus in Nebraska
}
