% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNV_models_lags.r
\name{models_lags}
\alias{models_lags}
\title{fits the models to training data and generates the predictions.}
\usage{
models_lags(models, allLagsT, allLagsO, results.path)
}
\arguments{
\item{models}{}

\item{allLagsT}{}

\item{allLagsO}{}

\item{results.path}{}
}
\description{
fits the models to training data and generates the predictions.
}
