% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm_NE_WNV.R
\docType{data}
\name{NEdat}
\alias{NEdat}
\title{NEdat}
\format{
An object of class \code{data.frame} with 23622 rows and 5 columns.
}
\source{
ftp://ftp.ncdc.noaa.gov/pub/data/cirs/climdiv/
}
\usage{
NEdat
}
\description{
temperature and precipitation data for Nebraska counties each month from
January 1998 to February 2019 from the National Centers for Environmental
Information, National Climatic Data Center
}
\keyword{datasets}
