% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R, R/predict_noYr.r,
%   R/predict_wYr.r
\name{cross_validate}
\alias{cross_validate}
\alias{predict_noYr}
\alias{predict}
\alias{predict_wYr}
\title{Predict data}
\usage{
cross_validate(fittedModel, allLagsT, fillzeros, allunits, nsim)

predict_noYr(fittedModel, allLagsT, allLagsO, fillzeros, allunits, nsim)

predict_wYr(fittedModel, allLagsT, allLagsO, fillzeros, allunits, nsim)
}
\arguments{
\item{fittedModel}{Model to predict from}

\item{allLagsT}{The assembled data set to use for training}

\item{fillzeros}{Logical. If TRUE fill in with zeros counties that have no cases ever.}

\item{allunits}{Character. Vector of county names in complete data set. Used 
when fillzeros == TRUE}

\item{nsim}{Integer. Number of samples to draw from posterior distribution. Defaults to zero, which has the expected value of cases in predcases.}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}
}
\description{
Predict data
}
\section{Functions}{
\itemize{
\item \code{cross_validate}: Generate leave-one-year-out cross-validated predictions from in-sample data

\item \code{predict_noYr}: Predict without year in model

\item \code{predict_wYr}: Predict from models including year
}}

