% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_noYr.r
\name{predict_noYr}
\alias{predict_noYr}
\title{Predict data without using year}
\usage{
predict_noYr(fittedModel, allLagsT, allLagsO)
}
\arguments{
\item{fittedModel}{Model to predict from}

\item{allLagsT}{The assembled data set to use for training}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}
}
\description{
Predict data without using year
}
