% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNV_models_lags.r
\name{predict_switch}
\alias{predict_switch}
\title{Parse the model formula and call the appropriate predict function}
\usage{
predict_switch(fittedmodel, allLagsT, allLagsO, fillzeros, allunits, nsim)
}
\arguments{
\item{allLagsT}{The assembled data set to use for training}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}

\item{fillzeros}{Logical. If TRUE fill in with zeros counties that have no cases ever.}

\item{allunits}{Character. Vector of county names in complete data set. Used 
when fillzeros == TRUE}

\item{nsim}{Integer. Number of samples to draw from posterior distribution. Defaults to zero, which has the expected value of cases in predcases.}

\item{fittedModel}{Model to predict from}
}
\value{

}
\description{
Parse the model formula and call the appropriate predict function
}
