% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wYr.r
\name{predict_wYr}
\alias{predict_wYr}
\title{Predict data without using year}
\usage{
predict_wYr(allLagsT, allLagsO)
}
\arguments{
\item{allLagsT}{The assembled data set to use for training}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}
}
\description{
Predict data without using year
}
