% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\name{run.flm}
\alias{run.flm}
\title{Run the Functional Linear Modeling approach using dfmip inputs}
\usage{
run.flm(human.data, weather.data, flm.inputs, weekinquestion, results.path,
  in.seed = 4872957)
}
\arguments{
\item{human.data}{#**# NEEDS DOCUMENTATION. human.data will be ignored if flm.inputs[[4]] is not NA}

\item{weather.data}{#**# NEEDS DOCUMENTATION, weather.data will be ignored if flm.inputs[[5]] is not NA.}

\item{flm.inputs}{Inputs specific to the functional linear modeling approach: \tabular{ll}{
pop \tab Human population for each county by year. Required fields are County, cofips, year, pop #**# is cofips needed?\cr
spi \tab Drought information corresponding to the spi drought index. #**# Update with link/chunk to other documentation\cr
spei \tab Drought information corresponding to the spei drought index #**# Update with link/chunk to other documentation\cr
NEdat \tab Weather data in format cofips, year, month, day, temp, ppt; where temp is mean temperature and ppt is mean precipitation.
If present, the weather.data argument will be ignored. If absent, this object will be created from the weather.data object.\cr}}

\item{weekinquestion}{The date for which the forecast is desired}

\item{results.path}{The path for outputting model results}

\item{in.seed}{A random seed to make the results repeatable}
}
\value{
flm.out A list of 3 objects: a results data frame, a data frame with random distribution draws, and a list of other outputs.
}
\description{
Run the Functional Linear Modeling approach using dfmip inputs
}
