% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_lags.R
\name{see.lags}
\alias{see.lags}
\title{See lags}
\usage{
see.lags(mod, results.path)
}
\arguments{
\item{mod}{A model object created by call.flm}

\item{results.path}{File path for output}
}
\description{
See lags
}
