% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\docType{data}
\name{spei}
\alias{spei}
\title{spei}
\format{
An object of class \code{data.frame} with 24552 rows and 4 columns.
}
\usage{
spei
}
\description{
Extracted monthly values from Westwide Drought Tracker netcdf files through
2019-12-15. See Abatzoglou, J. T., McEvoy, D. J., & Redmond, K. T. (2017).
The West Wide Drought Tracker: Drought monitoring at fine spatial scales.
Bulletin of the American Meteorological Society, 98(9), 1815–1820.
https://doi.org/10.1175/BAMS-D-16-0193.1
}
\keyword{datasets}
