% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\name{call.flm}
\alias{call.flm}
\title{Main function}
\usage{
call.flm(pop, cases, NEdat, spi, spei, target.date = "2018-02-01",
  start.year = 2002, results.path = "temp/", in.seed = 4872957)
}
\arguments{
\item{pop}{#**# Add documentation}

\item{cases}{#**# Add documentation}

\item{NEdat}{#**# Add documentation}

\item{spi}{#**# Add documentation}

\item{spei}{#**# Add documentation}

\item{target.date}{The last date to include for calculation of lags}

\item{start.year}{The first year to include in the training data}

\item{results.path}{The path where results will be written}

\item{in.seed}{The starting number for the random number generator. This makes the results repeatable.}
}
\description{
Main function
}
