% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNV_models_lags.r
\name{models_lags}
\alias{models_lags}
\title{Fits the models to training data and generates the predictions}
\usage{
models_lags(allmods, allLagsT, allLagsO, fillzeros, allunits, nsim)
}
\arguments{
\item{allmods}{List of models}

\item{allLagsT}{Training data}

\item{allLagsO}{Out-of-sample data}

\item{fillzeros}{Logical. If TRUE fill in with zeros counties that have no cases ever.}

\item{allunits}{Character. Vector of county names in complete data set. Used 
when fillzeros == TRUE}

\item{nsim}{Integer. Number of samples to draw from posterior distribution. Defaults to zero, which has the expected value of cases in predcases.}
}
\value{
Returns predictions for the out-of-sample year, a list of fitted models, AIC scores, and the model with the lowest AIC score
}
\description{
Fits the models to training data and generates the predictions
}
