% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.NE.WNV.R
\docType{data}
\name{NE_county_pops}
\alias{NE_county_pops}
\title{NE_county_pops}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1860 rows and 5 columns.
}
\usage{
NE_county_pops
}
\description{
Nebraska County populations, 2000-2019, from the U.S. Census Bureau, using annual estimates
}
\keyword{datasets}
