% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_noYr.r, R/predict_wYr.r
\name{predict_noYr}
\alias{predict_noYr}
\alias{predict}
\alias{predict_wYr}
\title{Predict data}
\usage{
predict_noYr(fittedModel, allLagsT, allLagsO, fillzeros, allunits, nsims)

predict_wYr(fittedModel, allLagsT, allLagsO, fillzeros, allunits, nsims)
}
\arguments{
\item{fittedModel}{Model to predict from}

\item{allLagsT}{The assembled data set to use for training}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}

\item{fillzeros}{Logical. If TRUE fill in with zeros counties that have no cases ever.}

\item{allunits}{Character. Vector of county names in complete data set. Used 
when fillzeros == TRUE}

\item{nsims}{Integer. Number of samples to draw from posterior distribution. Defaults to zero, which has the expected value of cases in predcases.}
}
\description{
Predict data
}
\section{Functions}{
\itemize{
\item \code{predict_noYr}: Predict without year in model

\item \code{predict_wYr}: Predict from models including year
}}

