% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wYr.r
\name{predict_wYr}
\alias{predict_wYr}
\title{Predict data, using imputation to get "new years"}
\usage{
predict_wYr(fittedModel, allLagsT, allLagsO)
}
\arguments{
\item{fittedModel}{Model to predict from}

\item{allLagsT}{The assembled data set to use for training}

\item{allLagsO}{The assembled data set to use for out-of-sample prediction}
}
\description{
Predict data, using imputation to get "new years"
}
