% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flm.R
\docType{data}
\name{sampledat}
\alias{sampledat}
\title{sampledat}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1488 rows and 4 columns.}
\usage{
sampledat
}
\description{
simulated data on annual numbers of human cases of neuro-invasive and non-neuro-invasive
West Nile Virus in Nebraska counties. It is predictions of a model that was trained on
actual numbers of cases as recorded in CDC's Arbonet database. It excludes Arthur County,
because no cases have been recorded there to date, and we had to exclude it from our
modeling to get it to work.
}
\keyword{datasets}
